requirements.txt
pandas-dq==1.29
streamlit==1.41.1
mysql-connector-python==8.3.0
matplotlib==3.9.2
numpy==1.23.5
plotly-express==0.4.1
sqlalchemy==2.0.34
altair==4.2.2
datetime
altair